Rename README.md to CFX Mining
<b>Mining pool Conflux</b>

As you know, cryptocurrency is a fairly significant coin in our realities, it costs more than a dollar and is quite stable - that is why mining well-known virtual banknotes has become a particularly popular hobby, such an activity can also be considered as a source of passive income. In order to mine as much crypto as possible, the network user should familiarize himself with the top sites with a pool - <a href="https://woolypooly.com/en/coin/cfx">CFX mining</a> has become an advanced development in this area.

What is a <a href="https://woolypooly.com/">mining pool</a>? Initially, every crypto owner who does not buy currency, but actually multiplies it, must have the appropriate equipment for such tasks, namely powerful video cards. Sometimes there is not enough capacity, and then a mining pool is used to extract coins - a single server that provides its own working capacities, uniting the efforts of a whole group of participants.
The task of the miners participating in the pool was to search for a block - as soon as a certain condition is met, the site issues a reward, which is distributed among those participating in the event. Such a service has truly become a godsend for the owners of one or another cryptocurrency, but it also has a negative impact - you can find many fraudulent sites, and on legal ones, the commission percentage is mandatory.

There are a great many people who want to mine and receive a block, thus creating fierce competition - that is why miners operating in a solo format need support, it is difficult for them to achieve effective results alone, and they are adjacent to a mining pool with a well-known name.

Today there is a large list of cryptocurrency names, the well-known bitcoin is of course relevant, but fresh solutions are also being released, for example, virtual Conflux - a profitable development with improved profitability indicators. On the site of the mining pool, you can receive a payment in Conflux, if required by the client of the system.

The topic of cryptocurrencies is very interesting for detailed consideration, but it is fraught with many nuances - fortunately, there are now many viewing platforms on the network where newcomers are helped to choose hardware and an operating strategy for quickly achieving goals, and material enrichment, of course.